Move utils (camelizeKeys) to core/api
